* moved sample grid component files under SampleGrid directory * added SampleGridContainer.js instead of SampleGridMain component * removed unused files * added SampleChanger routes file * new doGetSamplesList action, does a fetch on server to get samples list from SC * new queue and queue_model test hardware objects * add sample_changer to global app object
