+++ mxcube3/__init__.py
@@ -53,7 +53,8 @@ if not app.debug or os.environ.get("WERKZEUG_RUN_MAIN") == "true":
   app.db_connection = app.beamline.getObjectByRole("lims_client")
   #app.resolution = app.beamline.getObjectByRole("resolution")
   #app.queue = hwr.getHardwareObject(cmdline_options.queue_model)
+  app.sample_changer = app.beamline.getObjectByRole("sample_changer")
 
   ###Importing all REST-routes
-  import routes.Main, routes.Login, routes.Beamline, routes.Collection, routes.Mockups, routes.Sample, routes.SampleCentring
+  import routes.Main, routes.Login, routes.Beamline, routes.Collection, routes.Mockups, routes.Sample, routes.SampleCentring, routes.SampleChanger
