+++ mxcube3/ui/actions/samples_grid.js
@@ -1,5 +1,18 @@
+
+import fetch from 'isomorphic-fetch'
+
 export function doUpdateSamples(samples_list) {
     return { type: "UPDATE_SAMPLES", samples_list }
+}
+
+export function doGetSamplesList() {
+   return function(dispatch) {
+       fetch('mxcube/api/v0.1/sample_changer/samples_list')
+            .then(response => response.json())
+            .then(json => {
+                dispatch(doUpdateSamples(json.data));
+            })
+    }
 }
 
 export function doSetLoadable(loadable) {
