+++ mxcube3/ui/reducers/login.js
@@ -1,6 +1,7 @@
 const initialState = {
   data:{ }
 }
+
 export default (state=initialState, action) => {
     switch (action.type) {
         case 'LOGIN':
@@ -10,7 +11,7 @@ export default (state=initialState, action) => {
             window.error_notification.clear();
             return Object.assign({},state,initialState)
         default:
-            console.log("defalut")
             return state
     }
 }
+
