+++ test/HardwareObjectsMockup.xml/queue.xml
@@ -1,0 +1,4 @@
+<object class="QueueManager" role="Queue">
+  <object href="/beamline-setup" role="beamline_setup"/>    
+</object>
+
